# tripmatch_backend/docs/api-contract-iteratia-1-2.openapi.yaml
openapi: 3.0.3
info:
  title: TripMatch Backend API Contract (Iteratia 1 + 2 + 3)
  version: "iter1-iter2-iter3"
  description: >
    Contract extras direct din codebase (controller + DTO + validari service).
    Toate endpointurile protejate folosesc Basic Auth, exceptand /api/auth/**.
servers:
  - url: /
security:
  - basicAuth: []
...
paths:
...
  /api/accounts/{accountId}/profile:
    parameters:
      - $ref: "#/components/parameters/accountIdPath"
    get:
      tags: [Profile]
      summary: Get profile by accountId (owner only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Profile]
      summary: Update profile (owner or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
            example:
              phone: "+40712345678"
              city: "Cluj-Napoca"
              country: "Romania"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/trips/{tripId}/requests/publish:
    parameters:
      - $ref: "#/components/parameters/positiveTripIdPath"
    post:
      tags: [Request]
      summary: Publish or republish a travel request for a trip (user only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishTravelRequestRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRequestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/requests/{requestId}/unpublish:
    parameters:
      - $ref: "#/components/parameters/positiveRequestIdPath"
    patch:
      tags: [Request]
      summary: Unpublish (close) a travel request (user owner only)
      description: >
        Request status devine CLOSED.
        Ofertele cu status DRAFT/READY/SENT sunt marcate automat EXPIRED.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRequestResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/requests:
    get:
      tags: [Request]
      summary: List published travel requests for agencies (UC-11)
      parameters:
        - name: destination
          in: query
          required: false
          schema:
            type: string
            maxLength: 512
        - name: minBudget
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: maxBudget
          in: query
          required: false
          schema:
            type: number
            minimum: 0
        - name: startFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: flexibleDate
          in: query
          required: false
          schema:
            type: boolean
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamRequests"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTravelRequestListItemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/requests/{requestId}:
    parameters:
      - $ref: "#/components/parameters/positiveRequestIdPath"
    get:
      tags: [Request]
      summary: Get published request details for agency
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRequestDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/my/requests:
    get:
      tags: [Request]
      summary: List my travel requests (user only)
      parameters:
        - $ref: "#/components/parameters/pageParamDefault0"
        - $ref: "#/components/parameters/sizeParamDefault20"
        - $ref: "#/components/parameters/sortParamRequests"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTravelRequestListItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/my/requests/{requestId}:
    parameters:
      - $ref: "#/components/parameters/positiveRequestIdPath"
    get:
      tags: [Request]
      summary: Get my travel request details (user owner only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelRequestDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    pageParamDefault0:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    sizeParamDefault20:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
...
    sortParamRequests:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: "publishedAt,desc"
      description: Spring pageable sort, ex `publishedAt,desc`.
...
    accountIdPath:
      name: accountId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    positiveTripIdPath:
      name: tripId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    positiveRequestIdPath:
      name: requestId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: >
        Bad Request. In acest backend, body-ul poate fi:
        1) string simplu (Business error) sau
        2) array de stringuri (Bean validation errors).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationOrMessageError"
    Unauthorized:
      description: Missing/invalid Basic Auth credentials.
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: string
          example: "You cannot edit another user's profile"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: string
          example: "Trip not found!"
...
  schemas:
    ValidationOrMessageError:
      oneOf:
        - type: string
        - type: array
          items:
            type: string
...
    UpdateProfileRequest:
      type: object
      description: At least one field must be non-blank/non-null.
      properties:
        phone:
          type: string
          minLength: 5
          maxLength: 30
        profilePictureUrl:
          type: string
          maxLength: 500
        city:
          type: string
          maxLength: 50
        country:
          type: string
          maxLength: 50
        agencyDescription:
          type: string
          maxLength: 1000
    ProfileResponse:
      type: object
      properties:
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        profilePictureUrl: { type: string, nullable: true }
        city: { type: string, nullable: true }
        country: { type: string, nullable: true }
        agencyDescription: { type: string, nullable: true }
        agencyName: { type: string, nullable: true }
        isApproved: { type: boolean, nullable: true }
...
    TravelRequestStatus:
      type: string
      enum: [PUBLISHED, CLOSED]
...
    PublishTravelRequestRequest:
      type: object
      description: budgetMin <= budgetMax; endDate >= startDate.
      properties:
        sharedPreferences:
          type: string
          maxLength: 4000
          pattern: "^.*\\S.*$"
        budgetMin:
          type: number
          minimum: 0
        budgetMax:
          type: number
          minimum: 0
        destinationSummary:
          type: string
          maxLength: 512
          pattern: "^.*\\S.*$"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isFlexibleDate:
          type: boolean
      required: [sharedPreferences, budgetMin, budgetMax, destinationSummary, startDate, endDate, isFlexibleDate]
    TravelRequestResponse:
      type: object
      properties:
        requestId: { type: integer, format: int64 }
        tripId: { type: integer, format: int64 }
        status:
          $ref: "#/components/schemas/TravelRequestStatus"
        publishedAt: { type: string, format: date-time, nullable: true }
        message: { type: string }
      required: [requestId, tripId, status, message]
    TravelRequestListItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tripId: { type: integer, format: int64 }
        destinationSummary: { type: string }
        budgetMin: { type: number }
        budgetMax: { type: number }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        isFlexibleDate: { type: boolean }
        status:
          $ref: "#/components/schemas/TravelRequestStatus"
        publishedAt: { type: string, format: date-time }
    TravelRequestDetailResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tripId: { type: integer, format: int64 }
        userAccountId: { type: integer, format: int64 }
        sharedPreferences: { type: string }
        budgetMin: { type: number }
        budgetMax: { type: number }
        destinationSummary: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        isFlexibleDate: { type: boolean }
        status:
          $ref: "#/components/schemas/TravelRequestStatus"
        publishedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    PageTravelRequestListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TravelRequestListItemResponse"
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
      required: [content, page, size, totalElements, totalPages, last]
...